Null check for email report batch

+++ Source/SSCore/src/main/java/com/realtech/socialsurvey/core/services/reportingmanagement/impl/ReportingDashboardManagementImpl.java
@@ -5122,11 +5122,17 @@ public class ReportingDashboardManagementImpl<K> implements ReportingDashboardMa
 	@Transactional
 	public boolean saveEmailCountMonthData(List<SurveyInvitationEmailCountMonth> agentEmailCountsMonth) {
 		try {
+			int month = agentEmailCountsMonth.get(0).getMonth();
+			int year = agentEmailCountsMonth.get(0).getYear();
 			Map<String, Object> queryMap = new HashMap<String, Object>();
-			queryMap.put("month", agentEmailCountsMonth.get(0).getMonth());
-			queryMap.put("year", agentEmailCountsMonth.get(0).getYear());
-			surveyInvitationEmailDao.deleteOldDataForMonth(agentEmailCountsMonth.get(0).getMonth(),
-					agentEmailCountsMonth.get(0).getYear());
+			queryMap.put("month", month);
+			queryMap.put("year", year);
+			if(surveyInvitationEmailDao.getSurveyInvitationEmailReportCountForMonth(month, year) > 0) {
+				surveyInvitationEmailDao.deleteOldDataForMonth(agentEmailCountsMonth.get(0).getMonth(),
+						agentEmailCountsMonth.get(0).getYear());
+			} else {
+				LOG.info("No old data found to delete before save.");
+			}
 			surveyInvitationEmailDao.saveAll(agentEmailCountsMonth);
 			LOG.info("Survey invitaion email count data saved to db for {}-{}",agentEmailCountsMonth.get(0).getMonth(),
 					agentEmailCountsMonth.get(0).getYear());
